Moved search results into presentation
